# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Management API",
        "description": "API Documentation for Ticket Management System",
        "contact": {
            "name": "Abdulqdos Alabinie",
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
...
        "/api/v1/tickets/{event_id}/{ticketType_id}/store": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Book a ticket for an event",
                "description": "Book a ticket",
                "operationId": "08af301f4b661d6ea3aa3c8fa1be07a1",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketType_id",
                        "in": "path",
                        "description": "ID of the ticket type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket booked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Your Booked Has Create Check your profile"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation or business rule error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This ticket does not belong to the selected event"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{event_id}/{ticketType_id}/destroy": {
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Cancel a booked ticket",
                "description": "Cancel a booked ticket",
                "operationId": "f978355868a8f59f20005ac64e9cb630",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketType_id",
                        "in": "path",
                        "description": "ID of the ticket type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket canceled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Your Booked Has Deleted"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation or business rule error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not booked this ticket to cancel"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/ticket-types/{ticketType}": {
            "get": {
                "tags": [
                    "TicketTypes"
                ],
                "summary": "Get a single ticket type",
                "description": "Show single ticket type",
                "operationId": "aaaf0ee8ccfed966c5a03c5b72ea38e9",
                "parameters": [
                    {
                        "name": "ticketType",
                        "in": "path",
                        "description": "ID of the ticket type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket type details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket type not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket type not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TicketType": {
                "title": "TicketType",
                "description": "Ticket Type model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "event_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}